ENH: Add an option which allows the users to make a cross validation for the module
For this, in the tab 'Select Input Data', the user can specify if the input data was already in the groups used to create the Classification Groups. If this is the case, the classification groups are temporary recomputed without the input data vtk file to check if it belongs to the good group.